Correct README.md for Quad/Dual board
The description for the Quad/Dual board was pointing to the same branch used for
the Neo variant, but this branch does not feature support for the Quad/Dual board
any more, I've modified the README.md to point to the right branch and also fixed
the build instructions to the correct defconfig, i.e. udoo_qd_config instead of
udoo_qdl_config.

Signed-off-by: Manuel Rodriguez <manuel2982@gmail.com>